+++ geotrek/common/mixins/views.py
@@ -250,25 +250,23 @@ class BelongStructureMixin:
             return response
 
         # check permissions
-        queryset = self.get_queryset()
         user_structure = self.request.user.profile.structure
         has_bypass_structure_perm = self.request.user.has_perm(
             "authent.can_bypass_structure"
         )
 
-        filtered_queryset = queryset.filter(structure=user_structure)
+        has_wrong_structure_object = (
+            self.get_queryset().exclude(structure=user_structure).exists()
+        )
 
-        if (
-            not has_bypass_structure_perm
-            and filtered_queryset.count() != queryset.count()
-        ):
+        if not has_bypass_structure_perm and has_wrong_structure_object:
             messages.warning(
                 self.request,
                 _(
                     "Access is restricted because not all selected items belong to your structure. Use the structure filter to select only authorized items."
                 ),
             )
-            return HttpResponseRedirect(self.get_redirect_url())
+            return HttpResponseRedirect(self.get_success_url())
 
         return response