+++ geotrek/signage/tests/test_views.py
@@ -470,7 +470,6 @@ class BladeMultiActionsViewTest(
     model = Blade
     modelFactory = BladeFactory
     expected_fields = [
-        "Signage",
         "Direction",
         "Type",
         "Color",
@@ -478,15 +477,16 @@ class BladeMultiActionsViewTest(
 
     @classmethod
     def create_items(cls, struct=None):
-        signage1 = SignageFactory.create(structure=struct)
-        signage2 = SignageFactory.create(structure=StructureFactory.create())
-        cls.item1 = cls.modelFactory.create(signage=signage1)
-        cls.item2 = cls.modelFactory.create(signage=signage2)
+        cls.signage1 = SignageFactory.create(structure=struct)
+        cls.signage2 = SignageFactory.create(structure=StructureFactory.create())
+        cls.item1 = cls.modelFactory.create(signage=cls.signage1)
+        cls.item2 = cls.modelFactory.create(signage=cls.signage2)
 
-    def test_editable_fields_no_blades(self):
+    def test_editable_fields_no_blades_no_signage(self):
         self.client.force_login(self.user)
         response = self.client.get(
             self.model.get_multi_update_url() + f"?pks={self.item1.pk}"
         )
 
         self.assertNotContains(response, "Blades\n")
+        self.assertNotContains(response, "Signage\n")